+++ lib/open62541.dart
@@ -1,7 +1,8 @@
 library;
 
 export 'src/access_level.dart' show AccessLevelMask;
-export 'src/client.dart' show Client, ClientState, BrowseResultItem, BrowseTreeItem, NodeClass, BrowseResultMask;
+export 'src/client.dart' show Client, ClientState, DataValue, BrowseResultItem, BrowseTreeItem, NodeClass, BrowseResultMask;
+export 'src/common.dart' show statusCodeToString;
 export 'src/dynamic_value.dart' show DynamicValue, LocalizedText, EnumField, Schema;
 export 'src/extensions.dart'
     show
@@ -16,6 +17,10 @@ export 'src/extensions.dart'
 export 'src/third_party/open62541.g.dart'
     show
         UA_STATUSCODE_GOOD,
+        UA_STATUSCODE_BADNOCOMMUNICATION,
+        UA_STATUSCODE_BADNOTWRITABLE,
+        UA_STATUSCODE_BADUSERACCESSDENIED,
+        UA_STATUSCODE_BADINTERNALERROR,
         UA_OPEN62541_VER_MAJOR,
         UA_OPEN62541_VER_MINOR,
         UA_OPEN62541_VER_PATCH,
@@ -23,7 +28,7 @@ export 'src/third_party/open62541.g.dart'
         UA_OPEN62541_VER_COMMIT,
         UA_OPEN62541_VERSION;
 export 'src/node_id.dart' show NodeId;
-export 'src/server.dart' show Server, Argument;
+export 'src/server.dart' show Server, Argument, DataSourceValue;
 export 'src/types/errors.dart' show Inactivity, SecureChannelClosed, SubscriptionDeleted;
 export 'src/isolate.dart' show ClientIsolate;
 export 'src/client_api.dart' show ClientApi;